% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{as.peaklistSet.DataFrame}
\alias{as.peaklistSet.DataFrame}
\title{Force an peaklistSet Object to Belong to a Class DataFrame}
\usage{
as.peaklistSet.DataFrame(x, ...)
}
\arguments{
\item{...}{}
}
\value{
\code{\link[S4Vectors]{DataFrane} object}
}
\description{
Force an peaklistSet Object to Belong to a Class DataFrame
}
\examples{
library(rawDiag)
(rawfile <- file.path(path.package(package = 'rawDiag'), 'extdata', 'sample.raw'))
PLS <-readScans(rawfile)
DF <- as.peaklistSet.DataFrame(PLS)
DF$fromFile = as.integer(1)

if(require(Spectra)){
  rawDiagSample <- MsBackendDataFrame()
  BE <- backendInitialize(object=rawDiagSample, files=rawfile, spectraData=DF)
}
}
\author{
TK,CP 2019-06-14
}
