% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{as.rawDiag.mzR}
\alias{as.rawDiag.mzR}
\title{mzR reader method}
\usage{
as.rawDiag.mzR(object)
}
\arguments{
\item{object}{an mzR object}
}
\value{
an \code{\link{rawDiag}} object
}
\description{
mzR reader method
}
\examples{
if(require(mzR)){
   mzML <- "04_S174020_5000_5010.mzML"
   mzML <- file.path(path.package(package = "rawDiag"), "extdata", mzML)
   system.time(RAW <- rawDiag:::as.rawDiag.mzR(openMSfile(mzML)))
   summary(RAW)
   RAW$scanNumber
}

}
\author{
Christian Panse <cp@fgcz.ethz.ch>, Witold E.Wolski <wew@fgcz.ethz.ch>, 2018
}
