% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{as.rawDiag}
\alias{as.rawDiag}
\title{Title}
\usage{
as.rawDiag(object)
}
\arguments{
\item{object}{any R object.}
}
\value{
a \code{data.frame} having the column names defined in 
\code{rawDiag:::.getDesiredColNames}.
}
\description{
Title
}
