% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{benchmark_raw}
\alias{benchmark_raw}
\title{benchmark of \code{read.raw}}
\usage{
benchmark_raw(
  f,
  maxncpu = c(16, 32, 64),
  exe = file.path(path.package(package = "rawDiag"), "exec", "fgcz_raw.exe"),
  rdata = tempfile(fileext = ".RData")
)
}
\arguments{
\item{f}{set of files}

\item{maxncpu}{maximal number of cores used for the benchmark}

\item{exe}{the executable for reading the rawfiles}

\item{rdata}{file extrention of the tempfile}
}
\value{
a nested list object.
}
\description{
benchmarks the \code{read.raw} function's IO. This 
function has been used for generating \link{benchmark} data set.
}
\examples{
\dontrun{
 f <- f[grep("raw$",f<-list.files())]
 benchmark_raw(f, maxncpu = 8, exe='~/bin/fgcz_raw.exe', mono=TRUE, rdata='/tmp/b2.RData')
}
}
\references{
\url{http://planetorbitrap.com/rawfilereader#.WyThDK3QPmE}
}
\author{
Christian Panse <cp@fgcz.ethz.ch>, 2017, 2018
}
