% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{calc.cycle.time}
\alias{calc.cycle.time}
\title{Calculate MS Cycle Time}
\usage{
calc.cycle.time(x)
}
\arguments{
\item{x}{a \code{\link{data.frame}} fullfilling the \code{\link{is.rawDiag}} column naming criteria.}
}
\value{
calculates the time of all ms cycles and the 95% quantile value there of. 
the cycle time is defined as the time between two consecutive MS1 scans
}
\description{
Calculate MS Cycle Time
}
