% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{calc.master.scan}
\alias{calc.master.scan}
\title{Calculate Master Scan Number}
\usage{
calc.master.scan(x)
}
\arguments{
\item{x}{a data frame fullfilling the as.rawDiag column naming criteria}
}
\value{
calculates the MS1 master scan number of an MS2 scan and populates the MasterScanNumber
with it
}
\description{
Calculate Master Scan Number
}
