% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{checkRawrr}
\alias{checkRawrr}
\title{Checks Bioconductor installation instructions}
\usage{
checkRawrr()
}
\value{
TRUE if everything is installed correctly
}
\description{
Checks Bioconductor installation instructions
}
