% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{csReadScans}
\alias{csReadScans}
\title{readScans via rDotNet}
\usage{
csReadScans(rawfile, scans = NULL)
}
\arguments{
\item{rawfile}{the name of the Thermo Fisher Scietific raw file which the data
are to be read from.}

\item{scans}{a vector of requested scan numbers}
}
\value{
\CRANpkg{protViz} peaklist
}
\description{
readScans via rDotNet
}
\examples{
 (rawfile <- file.path(path.package(package = 'rawDiag'), 'extdata', 'sample.raw'))
 S <- csReadScans(rawfile, 1:9)
 plot(S[[1]])
 op <- par(mfrow=c(3, 3))
 lapply(S, function(x){plot(x, sub=x$scanType)})
}
