% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{.cycleTime}
\alias{.cycleTime}
\title{Calculate MS Cycle Time}
\usage{
.cycleTime(x)
}
\arguments{
\item{x}{a \code{\link{data.frame}} fullfilling the \code{is.rawDiag} column naming criteria.}
}
\value{
calculates the time of all ms cycles and the 95% quantile value there of. 
the cycle time is defined as the time between two consecutive MS1 scans
}
\description{
Graphs the lock mass deviations along RT.
}
\details{
TODO: quan
tile part needed? If no MS1 scan is present? -> DIA take lowest window as cycle indicator?
}
\author{
Christian Trachsel (2017), Christian Panse (20231201) refactored
}
