% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{.fillNAgaps}
\alias{.fillNAgaps}
\title{Fill NA values with last previous value}
\usage{
.fillNAgaps(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
a vector with any NA values replaced with the last previous actuall value
}
\description{
Fill NA values with last previous value
}
\examples{
c(NA, 1, 2, 3, NA, 4, 5, NA, NA, NA, 6) |>
  rawDiag:::.fillNAgaps()
}
\author{
Christian Trachsel
}
