% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XIC.R
\name{.XICextraction}
\alias{.XICextraction}
\title{extracting XIC of a given set of peptides}
\usage{
.XICextraction(
  rawfilename = "/Users/cp/data//Data_small/F3_20170823_02_autoQC01_threshold.raw",
  pepSeq = c("AGGSSEPVTGLADK", "GDLDAASYYAPVR", "FLLQFGAQGSPLFK"),
  debug = FALSE
)
}
\arguments{
\item{rawfilename}{rawfile}

\item{pepSeq}{Amino acid sequence (peptide)}
}
\value{
peptide quant information
}
\description{
ALLES wird gut
}
\author{
CP/BR
}
