% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promega.R
\name{getPromega6x5mix}
\alias{getPromega6x5mix}
\alias{extractPromegaXIC}
\title{Get in-silico promega's 6 x 5 lc ms ms peptide reference mix}
\usage{
getPromega6x5mix()
}
\value{
a nested list containing the in-silico peptide information
}
\description{
Get in-silico promega's 6 x 5 lc ms ms peptide reference mix
}
\details{
contains the data contained in Table 5. 
'Masses of Various Peptides. Underlined letters indicate peptides labeled
with stable, heavy isotopes'.
}
\examples{
promega <- getPromega6x5mix()
\dontrun{
start <- Sys.time()
filepathroot <- '/Users/cp/Downloads/'
rawfile <- file.path(filepathroot, "20181012_002_S186065_6x5LC-MS2_A_Method_1.raw")

mZ <- sapply(promega, function(x){ (parentIonMass(x$sequence) + sum(x$mod) + 2 * 1.00794) / 2 })
abundance <- sapply(promega, function(x){x$abundance})
sequence <- sapply(promega, function(x){x$sequence})

XIC <- readXICs(rawfile, mZ, tol = 10)
XIC.log.max <- sapply(XIC, function(x){log(max(x$intensities))})

library(lattice)
xyplot(XIC.log.max ~ log(abundance) | sequence,
       panel = function(x, y, ...){panel.xyplot(x,y,... ); 
           fm <- lm(y ~ x); panel.lines(x, fitted(fm))}, pch=16, main=basename(rawfile),
       sub=list(paste("analyse the promega 6-x-5-lc-ms-ms-peptide-reference-mix-protocol
                      with rawDiag in less than 10 lines of R code in less than", 
                      round(Sys.time() - start) +1, "seconds."), cex=1))
}
}
\references{
\href{https://www.promega.com/-/media/files/resources/protocols/technical-manuals/101/6-x-5-lc-ms-ms-peptide-reference-mix-protocol.pdf}{6-x-5-lc-ms-ms-peptide-reference-mix-protocol}
}
\author{
Tobias Kockmann and Christian Panse, 2018
}
