% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{getWU163763}
\alias{getWU163763}
\title{WU163763}
\usage{
getWU163763()
}
\value{
a \code{\link{data.frame}} fullfilling the \code{\link{is.rawDiag}} column naming criteria.
}
\description{
a data set generated from reading \url{https://fgcz-bfabric.uzh.ch/bfabric/userlab/show-workunit.html?id=163763}.
}
