% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{is.rawDiag}
\alias{is.rawDiag}
\title{Is an Object an rawDiag Object?}
\usage{
is.rawDiag(object)
}
\arguments{
\item{object}{any \R object.}
}
\value{
a boolean
}
\description{
Is an Object an rawDiag Object?
}
\examples{
rawrr::sampleFilePath() |> rawDiag::readRaw() |> rawDiag::is.rawDiag()

}
\author{
Christian Panse 2018
}
