% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{mass.distribution}
\alias{mass.distribution}
\title{mass distribution plot}
\usage{
mass.distribution(x)
}
\arguments{
\item{x}{a data.frame object as returned by \code{\link{read.raw}}}
}
\value{
a ggplot object
}
\description{
plots the mass frequency in dependency to the charge state
}
\examples{

 load(file.path(path.package(package = "rawDiag"),
  file.path("extdata", "WU163763.RData")))
 
 mass.distribution(WU163763) +
   facet_wrap(~filename, ncol = 3)
 
 rawDiag:::mass.distribution.overlay(WU163763) +
   theme(legend.position = 'none')
   
 rawDiag:::mass.distribution.violin(WU163763) +
   theme(legend.position = 'none')  
}
