% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{mass.heatmap}
\alias{mass.heatmap}
\title{mass heatmap}
\usage{
mass.heatmap(x)
}
\arguments{
\item{x}{an rawDiag object}
}
\value{
a gglot object
}
\description{
graphs a deconvoluted heatmap of the StartTime
}
