% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgf.R
\name{mgf}
\alias{mgf}
\title{generate a mascot generic file}
\usage{
mgf(rawfilename, mgffilename = paste(sub("\\\\.raw$", "", rawfilename),
  "mgf", sep = "."))
}
\arguments{
\item{rawfilename}{}

\item{mgffilename}{}
}
\description{
generate a mascot generic file
}
