% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{plot.peaklist}
\alias{plot.peaklist}
\title{plot a peaklist}
\usage{
plot.peaklist(x, y, ...)
}
\arguments{
\item{x}{a peaklist}

\item{y}{}

\item{...}{}
}
\value{
a nested list
}
\description{
plot a peaklist
}
\examples{
(rawfile <- file.path(path.package(package = 'rawDiag'), 'extdata', 'sample.raw'))
S <- readScans(rawfile, 1:10)
plot(S[[1]])

}
