% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{plot.XICs}
\alias{plot.XICs}
\title{plot extracted ion chromatograms}
\usage{
plot.XICs(x, y, method = "ggplot", ...)
}
\arguments{
\item{x}{a list of XIC objects}

\item{y}{will be ignroed}

\item{...}{passed to \code{plot.XIC}}
}
\description{
plot extracted ion chromatograms
}
\examples{
# Example 1: extract iRT peptides
iRTpeptide <- c("LGGNEQVTR", "YILAGVENSK", "GTFIIDPGGVIR", "GTFIIDPAAVIR",
"GAGSSEPVTGLDAK", "TPVISGGPYEYR", "VEATFGVDESNAK",
"TPVITGAPYEYR", "DGLDAASYYAPVR", "ADVTPADFSEWSK",
"LFLQFGAQGSPFLK")

library(protViz)
# 2Hplus 
(mZ <- (parentIonMass(iRTpeptide) + 1.008) / 2)

\dontrun{
 library(ggplot2)
 rawfile <- "/home/cp/Downloads/20180220_14_autoQC01.raw"
 X <-readXICs(rawfile, masses=mZ)
 plot(X, method='ggplot') + facet_wrap(~mass)
 plot(X, method='ggplot') + facet_wrap(~mass, scales = "free_x")
 
}
}
