% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{PlotAll}
\alias{PlotAll}
\title{PlotAll applies all available rawDiag plot functions to a given rawDiag object}
\usage{
PlotAll(x, prefix = "primer", savepng = TRUE, resolution = 240)
}
\arguments{
\item{x}{a \code{\link{data.frame}} fullfilling the \code{\link{is.rawDiag}} column naming criteria.}

\item{prefix}{filename prefix if plots a writting to a png file}

\item{savepng}{default is TRUE}

\item{resolution}{number of pixels in png file, default is 240}
}
\description{
PlotAll applies all available rawDiag plot functions to a given rawDiag object
}
\examples{
\dontrun{
data(WU163763)
WU <- WU163763[WU163763$filename \%in\% unique(WU163763$filename)[1:2], ]
rv <- rawDiag:::PlotAll(x = WU, savepng = FALSE)
}
}
\references{
Table 1 in \url{https://doi.org/10.1101/304485}
}
\author{
Christian Panse <cp@fgcz.ethz.ch>, 2018
}
