% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{plotCycleTime}
\alias{plotCycleTime}
\title{Plot Cycle Time}
\usage{
plotCycleTime(x, method = "trellis")
}
\arguments{
\item{x}{a \code{\link{data.frame}} fullfilling the \code{is.rawDiag} column naming criteria.}

\item{method}{a character 'trellis', 'violin' or 'overlay'.}
}
\value{
a \code{\link{ggplot2}} object.
}
\description{
Graphs the time difference between two consecutive MS1 scans
(cycle time) with respect to RT (scatter plots) or its density (violin).
A smooth curve graphs the trend. The 95th percentile is indicated by a red
dashed line.
}
