% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{PlotInjectionTime}
\alias{PlotInjectionTime}
\alias{injection.time.overlay}
\title{injection.time}
\usage{
PlotInjectionTime(x, method = "trellis")
}
\arguments{
\item{x}{a \code{\link{data.frame}} fullfilling the \code{\link{is.rawDiag}} column naming criteria.}

\item{method}{a character 'trellis' or 'overlay'.}
}
\value{
ggplot object
}
\description{
injection.time
}
