% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{PlotMzDistribution}
\alias{PlotMzDistribution}
\alias{mz.distribution.overlay}
\alias{mz.distribution.violin}
\title{mz distribution}
\usage{
PlotMzDistribution(x, method = "trellis")
}
\arguments{
\item{x}{a \code{\link{data.frame}} fullfilling the \code{\link{is.rawDiag}} column naming criteria.}

\item{method}{a character 'trellis', 'violin' or 'overlay'.}
}
\value{
a ggplot objecy
}
\description{
mz distribution
}
