% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{PlotPrecursorHeatmap}
\alias{PlotPrecursorHeatmap}
\title{PrecursorMass versus StartTime hexagons MS2}
\usage{
PlotPrecursorHeatmap(x, method = "overlay", bins = 80)
}
\arguments{
\item{x}{a \code{\link{data.frame}} fullfilling the \code{\link{is.rawDiag}} column naming criteria.}

\item{method}{plot method.}

\item{bins}{number of bins in both vertical and horizontal directions. default is 80.}
}
\value{
a ggplot object.
}
\description{
PrecursorMass versus StartTime hexagons MS2
}
\note{
TODO: define bin with dynamically as h= 2x IQR x n e-1/3 or number of bins (max-min)/h
}
