% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{plotTicBasepeak}
\alias{plotTicBasepeak}
\title{TIC and Base Peak plot function}
\usage{
plotTicBasepeak(x, method = "trellis")
}
\arguments{
\item{x}{a \code{\link{data.frame}} fullfilling the \code{is.rawDiag} column naming criteria.}

\item{method}{a character 'trellis', 'violin' or 'overlay'.}
}
\value{
a ggplot object for graphing the TIC and the Base Peak chromatogram
}
\description{
Function for displaying the Total Ion Cound (TIC) and Base
Peak chromatogram of a mass spectrometry measurement. 
Multiple files are handled by faceting based on rawfile name.
}
