% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{precursor.heatmap}
\alias{precursor.heatmap}
\title{PrecursorMass versus StartTime hexagons MS2}
\usage{
precursor.heatmap(x, ...)
}
\arguments{
\item{x}{an rawDiag object}
}
\value{
a ggplot object
}
\description{
PrecursorMass versus StartTime hexagons MS2
}
