% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\docType{package}
\name{rawDiag-package}
\alias{rawDiag-package}
\title{rawDiag: Brings Orbitrap Mass Spectrometry Data to Life; Fast and Colorful}
\description{
Optimizing methods for liquid chromatography coupled to mass spectrometry (LC-MS) poses a nontrivial challenge. The rawDiag package facilitates rational method optimization by generating MS operator-tailored diagnostic plots of scan-level metadata. The package is designed for use on the R shell or as a Shiny application on the Orbitrap instrument PC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cpanse/rawDiag/}
  \item Report bugs at \url{https://github.com/cpanse/rawDiag/issues}
}

}
\author{
\strong{Maintainer}: Christian Panse \email{cp@fgcz.ethz.ch} (\href{https://orcid.org/0000-0003-1975-3064}{ORCID})

Authors:
\itemize{
  \item Christian Trachsel \email{Christian.Trachsel@fgcz.ethz.ch}
  \item Tobias Kockmann \email{Tobias.Kockmann@fgcz.ethz.ch}
}

}
\keyword{internal}
