% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-rawDiag.R
\name{rawDiagServer}
\alias{rawDiagServer}
\title{rawDiag shiny module}
\usage{
rawDiagServer(id, vals)
}
\arguments{
\item{id}{the shiny session id.}

\item{vals}{reactiveValues containing, e.g., the rawfile path.}
}
\description{
rawDiag shiny module
}
\author{
Christian Panse (2018-2023)
}
