% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{rawDiagShiny}
\alias{rawDiagShiny}
\title{run the package demo}
\usage{
rawDiagShiny(
  appDir = system.file("shiny", "demo", package = "rawDiag"),
  root = Sys.getenv("HOME"),
  dirlist = unique(dirname(list.files(root, recursive = TRUE, pattern = "*.raw"))),
  ...
)
}
\arguments{
\item{appDir}{directory containing the shiny \code{server.R},}

\item{root}{default is the home directory.}

\item{dirlist}{all directories containg a raw files.}

\item{...}{passed to the \code{\link{runApp}} method.}
}
\value{
\code{NULL}
}
\description{
run the package demo
}
\examples{
\dontrun{
rawDiagShiny(root="/Users/cp/Downloads", 
  dirlist=unique(dirname(list.files("/Users/cp/Downloads", recursive = TRUE, pattern="*.raw"))))

## run it from the command line

# MacOSX and Linux
R -e "library(rawDiag);rawDiagShiny(launch.browser=TRUE)"

# Windows
R.exe -e "library(rawDiag);rawDiagShiny(root='D:/Data2Sam/', launch.browser=TRUE)"
}
}
