% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{rawDiagUI}
\alias{rawDiagUI}
\title{rawDiag shiny module UI}
\usage{
rawDiagUI(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
a shiny UI module

rawDiag shiny module UI
}
\description{
rawDiag shiny module UI
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rawDiag::shiny(rawDir = (rawrr::sampleFilePath() |> dirname()))
\dontshow{\}) # examplesIf}
}
