% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{read.raw.info}
\alias{read.raw.info}
\title{Read raw file meta data}
\usage{
read.raw.info(file, mono = if (Sys.info()["sysname"] \%in\% c("Darwin",
  "Linux")) TRUE else FALSE, exe = file.path(path.package(package =
  "rawDiag"), "exec", "fgcz_raw.exe"), mono_path = "", argv = "info",
  system2_call = TRUE, method = "thermo")
}
\arguments{
\item{file}{Thermo Fisher raw file name}

\item{mono}{enviroment}

\item{exe}{path of the executable.}

\item{mono_path}{default.}

\item{argv}{arguments, default.}

\item{system2_call}{system2 call, default.}

\item{method}{instrument vendor}
}
\value{
a nested list object
}
\description{
The function extracts some meta information from a given rawfile.
The output is parsed by the yaml package.
}
\examples{
(rawfile <- file.path(path.package(package = 'rawDiag'), 'extdata', 'sample.raw'))
Y <- read.raw.info(rawfile)
}
\seealso{
Thermo Fisher NewRawfileReader C# code snippet.
}
\author{
Christian Panse 2018, 2019
}
