% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{read.raw}
\alias{read.raw}
\title{reads selected raw file trailer for rawDiag plot functions}
\usage{
read.raw(rawfile)
}
\arguments{
\item{rawfile}{the name of the raw file containing the mass
spectrometry data from the Thermo Fisher Scientific instrument.}
}
\description{
reads selected raw file trailer for rawDiag plot functions
}
\author{
Christian Panse (2016-2023)
}
