% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{read.tdf}
\alias{read.tdf}
\title{Reading Bruker tdf files}
\usage{
read.tdf(filename)
}
\arguments{
\item{filename}{filename if the tdf file}
}
\value{
a rawDiagobject
}
\description{
Reading Bruker tdf files
}
\details{
Id LargestPeakMz AverageMz MonoisotopicMz Charge ScanNumber Intensity Parent
1  1      827.5638  828.0039       827.5638      1   623.2419      3239      3
2  2      727.6347  727.9152       727.6347      1   682.1215      2610      3
}
\note{
this is work in progress
}
\author{
Witold E. Wolski, Christian Panse <wew,cp@fgcz.ethz.ch>, 2018
}
