% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{readScans}
\alias{readScans}
\title{readScans of a raw file}
\usage{
readScans(rawfile, scans = NULL)
}
\arguments{
\item{rawfile}{the name of the Thermo Fisher Scietific raw file which the data
are to be read from.}

\item{scans}{a vector of requested scan numbers.}
}
\value{
\CRANpkg{protViz} peaklist
}
\description{
the function reads scan information of a given set of scan
number using a dot net interface and the ThermoFisher NewRawFileReader 
libraries.
}
\note{
the deprecated function is available through 
using \code{rawDiag:::.readScans}.
}
\references{
\url{https://doi.org/10.5281/zenodo.2640013}
}
\seealso{
\url{https://CRAN.R-project.org/package=rDotNet}
}
\author{
Christian Panse <cp@fgz.ethz.ch> 2018, 2019
}
