% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{scan.frequ.moving.aver}
\alias{scan.frequ.moving.aver}
\title{Calculate moving average of scan frequency}
\usage{
scan.frequ.moving.aver(x)
}
\arguments{
\item{x}{a data frame fullfilling the as.rawDiag column naming criteria}
}
\description{
Calculate moving average of scan frequency
}
