% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{scan.frequency}
\alias{scan.frequency}
\title{Title}
\usage{
scan.frequency(x)
}
\arguments{
\item{x}{a dataframe passing the is.rawDiag function}
}
\value{
a trellis like plot object displaying the scan frequency of a mass spec run
}
\description{
Title
}
