% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{ScanFrequMovingOver}
\alias{ScanFrequMovingOver}
\title{Calculate moving average of scan frequency}
\usage{
ScanFrequMovingOver(x)
}
\arguments{
\item{x}{a \code{\link{data.frame}} fullfilling the \code{\link{is.rawDiag}} column naming criteria.}
}
\value{
a \code{data.frame}.
}
\description{
Calculate moving average of scan frequency
}
