% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-rawDiag.R
\name{shiny}
\alias{shiny}
\title{Run the rawDiag shiny application}
\usage{
shiny(
  appDir = system.file("shiny", package = "rawDiag"),
  rawDir = (file.path(Sys.getenv("HOME"), "Downloads")),
  ...
)
}
\arguments{
\item{appDir}{The application to run. Should be one of the following:
\itemize{
\item A directory containing \code{server.R}, plus, either \code{ui.R} or
a \code{www} directory that contains the file \code{index.html}.
\item A directory containing \code{app.R}.
\item An \code{.R} file containing a Shiny application, ending with an
expression that produces a Shiny app object.
\item A list with \code{ui} and \code{server} components.
\item A Shiny app object created by \code{\link[shiny:shinyApp]{shinyApp()}}.
}}

\item{rawDir}{A directory containing the input raw files,
default is set to the \code{$HOME/Downloads} directory.}

\item{\ldots}{passed to the \code{\link{runApp}} method.}
}
\description{
Run the rawDiag shiny application
}
\examples{
\dontrun{
rawDiag::shiny(launch.browser = TRUE, display.mode = 'normal')

## embracing the command line

# MacOSX and Linux
R -q -e "rawDiag::shiny(launch.browser = TRUE)"

# Microsoft Windows
R.exe -e "rawDiag::shiny(launch.browser = TRUE)"
}
}
