% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{tic.basepeak}
\alias{tic.basepeak}
\title{TIC and Base Peak plot function}
\usage{
tic.basepeak(x)
}
\arguments{
\item{x}{a data frame fullfilling the as.rawDiag column naming criteria}
}
\value{
a ggplot object for graphing the TIC and the Base Peak chromatogram
}
\description{
Function for displaying the TIC and Base Peak chromatogram of a mass spectrometry measurement. 
Multiple files are handled by faceting based on filename.
}
