% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{plot.XIC}
\alias{plot.XIC}
\title{plot extracted ion chromatogram}
\usage{
plot.XIC(x, y, fit = FALSE, ...)
}
\arguments{
\item{x}{an XIC S3 class object}

\item{y}{will be ignored}

\item{fit}{\code{TRUE} or \code{FALSE}, if a lm fit should be done}

\item{...}{passed to the plot function}
}
\value{
plot or ggplot object
}
\description{
the defaukt plot function for an XIC object.
}
