% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{PlotCycleLoad}
\alias{PlotCycleLoad}
\alias{cycle.load.overlay}
\alias{cycle.load.violin}
\title{cycle.load plot}
\usage{
PlotCycleLoad(x, method = "trellis")
}
\arguments{
\item{x}{a \code{\link{data.frame}} fullfilling the \code{\link{is.rawDiag}} column naming criteria.}

\item{method}{a character 'trellis', 'violin' or 'overlay'.}
}
\value{
a ggplot object.
}
\description{
cycle.load plot
}
