% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{PlotCycleTime}
\alias{PlotCycleTime}
\alias{cycle.time.violin}
\alias{cycle.time.overlay}
\title{cycle time}
\usage{
PlotCycleTime(x, method = "trellis")
}
\arguments{
\item{x}{a \code{\link{data.frame}} fullfilling the \code{\link{is.rawDiag}} column naming criteria.}

\item{method}{a character 'trellis', 'violin' or 'overlay'.}
}
\value{
a \code{\link{ggplot}} object.
}
\description{
graphs cycle time versus rt. each item represents the time for one scan cycle.
}
