% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{PlotCycleTime}
\alias{PlotCycleTime}
\title{Plot Cycle Time}
\usage{
PlotCycleTime(x, method = "trellis")
}
\arguments{
\item{x}{a \code{data.frame} object adhering to the specified criteria for
the \code{is.rawDiag} function.}

\item{method}{specifying the plot method 'trellis' | 'violin' | 'overlay'.
The default is 'trellis'.}
}
\value{
a \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot Cycle Time
}
