% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{plotCycleTime}
\alias{plotCycleTime}
\title{cycle time plot}
\usage{
plotCycleTime(x, method = "trellis")
}
\arguments{
\item{x}{a \code{\link{data.frame}} fullfilling the \code{is.rawDiag} column naming criteria.}

\item{method}{a character 'trellis', 'violin' or 'overlay'.}
}
\value{
a \code{\link{ggplot2}} object.
}
\description{
graphs cycle time versus rt.
each item represents the time for one scan cycle.
}
