% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{plotInjectionTime}
\alias{plotInjectionTime}
\title{Plot Injection Time}
\usage{
plotInjectionTime(x, method = "trellis")
}
\arguments{
\item{x}{a \code{data.frame} object adhering to the specified criteria for the \code{is.rawDiag} function.}

\item{method}{specifying the plot method 'trellis' | 'violin' | 'overlay'. The default is 'trellis'.}
}
\value{
a \code{\link{ggplot2}} object.
}
\description{
Plot Injection Time
}
\examples{
rawrr::sampleFilePath() |> read.raw() |> plotInjectionTime()
}
