% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{PlotLockMassCorrection}
\alias{PlotLockMassCorrection}
\alias{lm.correction.violin}
\alias{lm.correction.overlay}
\title{lock mass correction plot}
\usage{
PlotLockMassCorrection(x, method = "trellis")
}
\arguments{
\item{x}{a \code{\link{data.frame}} fullfilling the \code{\link{is.rawDiag}} column naming criteria.}

\item{method}{a character 'trellis', 'violin' or 'overlay'.}
}
\value{
a ggplot object.
}
\description{
lock mass correction plot
}
