% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{plotLockMassCorrection}
\alias{plotLockMassCorrection}
\title{lock mass correction plot}
\usage{
plotLockMassCorrection(x, method = "trellis")
}
\arguments{
\item{x}{a \code{\link{data.frame}} fullfilling the \code{is.rawDiag} column naming criteria.}

\item{method}{a character 'trellis', 'violin' or 'overlay'.}
}
\value{
a ggplot2 object
}
\description{
lock mass correction plot
}
\examples{
rawrr::sampleFilePath() |>
  read.raw() |>
  plotLockMassCorrection()
}
\references{
rawDiag \doi{10.1021/acs.jproteome.8b00173}
}
\author{
Christian Trachsel (2017), Christian Panse (2023)
}
