% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{PlotMassDistribution}
\alias{PlotMassDistribution}
\alias{mass.distribution.violin}
\alias{mass.distribution.overlay}
\title{mass distribution plot}
\usage{
PlotMassDistribution(x, method = "trellis")
}
\arguments{
\item{x}{a \code{\link{data.frame}} fullfilling the \code{\link{is.rawDiag}} column naming criteria.}

\item{method}{a character 'trellis', 'violin' or 'overlay'.}
}
\value{
a ggplot object.
}
\description{
plots the mass frequency in dependency to the charge state
}
\examples{
 library(ggplot2)
 data(WU163763)
 
 PlotMassDistribution(WU163763, method = 'trellis') +
   facet_wrap(~ filename, ncol = 3)
 
 PlotMassDistribution(WU163763, method = 'violin') +
   theme(legend.position = 'none')
   
 PlotMassDistribution(WU163763, method = 'overlay') +
   theme(legend.position = 'none')  
}
