% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{plotPrecursorHeatmap}
\alias{plotPrecursorHeatmap}
\title{precursorMass versus StartTime hexagons MS2}
\usage{
plotPrecursorHeatmap(x, method = "overlay", bins = 80)
}
\arguments{
\item{x}{a \code{\link{data.frame}} fullfilling the \code{is.rawDiag} column naming criteria.}

\item{method}{a character 'trellis', 'violin' or 'overlay'.}

\item{bins}{number of bins in both vertical and horizontal directions. default is 80.}
}
\value{
a ggplot2 object.
}
\description{
precursorMass versus StartTime hexagons MS2
}
\note{
TODO: define bin with dynamically as h= 2x IQR x n e-1/3 or number of bins (max-min)/h
}
\author{
Christian Trachsel (2017)
}
