% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{PlotScanTime}
\alias{PlotScanTime}
\alias{scan.time.overlay}
\alias{scan.time.violin}
\title{scan event}
\usage{
PlotScanTime(x, method = "trellis")
}
\arguments{
\item{x}{a \code{\link{data.frame}} fullfilling the \code{\link{is.rawDiag}} column naming criteria.}

\item{method}{a character 'trellis', 'violin' or 'overlay'.}
}
\value{
a ggplot object.
}
\description{
plots time for each scan event
}
