% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{PlotTicBasepeak}
\alias{PlotTicBasepeak}
\title{Total Ion Count and Base Peak Plot}
\usage{
PlotTicBasepeak(x, method = "trellis")
}
\arguments{
\item{x}{a \code{data.frame} object adhering to the specified criteria for
the \code{is.rawDiag} function.}

\item{method}{specifying the plot method 'trellis' | 'violin' | 'overlay'.
The default is 'trellis'.}
}
\value{
a ggplot2 object for graphing the TIC and the Base Peak chromatogram.
}
\description{
Total Ion Count and Base Peak Plot
}
\author{
Christian Trachsel (2017), Christian Panse (20231130) refactored
}
