% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-rawDiag.R
\name{rawDiagServer}
\alias{rawDiagServer}
\title{rawDiag shiny module}
\usage{
rawDiagServer(id, vals)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{vals}{containing rawfile}
}
\description{
rawDiag shiny module
}
