% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-rawDiag.R
\name{rawDiagUI}
\alias{rawDiagUI}
\title{rawDiag shiny module UI}
\usage{
rawDiagUI(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\description{
rawDiag shiny module UI
}
