% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{read.raw}
\alias{read.raw}
\title{mass spec reader function}
\usage{
read.raw(file, mono = FALSE, exe = file.path(path.package(package =
  "rawDiag"), "exec/fgcz_raw.exe"), argv = "qc", method = "thermo",
  ssh = FALSE)
}
\arguments{
\item{file}{the name of the Thermo Fisher Scietific raw file which the data
are to be read from.}

\item{mono}{a boolean indicates if the mono enviroment should be used.}

\item{exe}{the adapter software.}

\item{argv}{argument vector for the adapter.}

\item{method}{the adapter type. default is 'thermo' which read Thermo Fisher Scietific raw files.}

\item{ssh}{boolean, if ssh pipe should be used. defaul is FALSE
@author Christian Panse <cp@fgcz.ethz.ch>, 2018}
}
\value{
a \code{data.frame}.
}
\description{
this function is a generic adapter function for 
  reading mass spectrometric measurement.
}
\examples{
read.raw
}
\seealso{
\url{http://planetorbitrap.com/rawfilereader}
}
