% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{readScans}
\alias{readScans}
\title{read scan of scanids}
\usage{
readScans(rawfile, scans, mono = if (Sys.info()["sysname"] \%in\% c("Darwin",
  "Linux")) TRUE else FALSE, exe = file.path(path.package(package =
  "rawDiag"), "exec", "fgcz_raw.exe"))
}
\arguments{
\item{rawfile}{}

\item{scans}{}

\item{mono}{}

\item{exe}{}
}
\description{
read scan of scanids
}
\examples{
 (rawfile <- file.path(path.package(package = 'rawDiag'), 'extdata', 'sample.raw'))
 S <- readScans(rawfile, 1:9)
 plot(S[[1]])
 op <- par(mfrow=c(3, 3))
 lapply(S, function(x){plot(x, sub=x$scanType)})

}
