% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{readXICs}
\alias{readXICs}
\title{Extracts XICs of a given mass vector}
\usage{
readXICs(
  rawfile,
  masses,
  tol = 10,
  mono = if (Sys.info()["sysname"] \%in\% c("Darwin", "Linux")) TRUE else FALSE,
  exe = file.path(path.package(package = "rawDiag"), "exec", "fgcz_raw.exe")
)
}
\arguments{
\item{rawfile}{the file name}

\item{masses}{a vector of masses}

\item{tol}{tolerance in ppm}

\item{mono}{if the mono enviroment should be used.}

\item{exe}{the exe file user by mono}
}
\value{
list of XIC objects
}
\description{
Extracts XICs of a given mass vector
}
\examples{
# Example 1: extract iRT peptides
iRTpeptide <- c("LGGNEQVTR", "YILAGVENSK", "GTFIIDPGGVIR", "GTFIIDPAAVIR",
"GAGSSEPVTGLDAK", "TPVISGGPYEYR", "VEATFGVDESNAK",
"TPVITGAPYEYR", "DGLDAASYYAPVR", "ADVTPADFSEWSK",
"LFLQFGAQGSPFLK")

library(protViz)
# 2Hplus 
(mZ <- (parentIonMass(iRTpeptide) + 1.008) / 2)

\dontrun{
rawfile <- "/home/cp/Downloads/20180220_14_autoQC01.raw"
 X <-readXICs(rawfile, masses=mZ)
}


(rawfile <- file.path(path.package(package = 'rawDiag'), 'extdata', 'sample.raw'))
RAW <- read.raw(rawfile)

# not meaning full but proof-of-concept
X <- readXICs(rawfile, masses=unique(RAW$PrecursorMass), tol=1000)
plot.XIC(X)

}
