% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-rawDiag.R
\name{shiny}
\alias{shiny}
\title{Run the rawDiag shiny application}
\usage{
shiny(
  appDir = system.file("shiny", package = "rawDiag"),
  rawDir = (file.path(Sys.getenv("HOME"), "Downloads")),
  ...
)
}
\arguments{
\item{appDir}{The application to run. Should be one of the following:
\itemize{
\item A directory containing \code{server.R}, plus, either \code{ui.R} or
a \code{www} directory that contains the file \code{index.html}.
\item A directory containing \code{app.R}.
\item An \code{.R} file containing a Shiny application, ending with an
expression that produces a Shiny app object.
\item A list with \code{ui} and \code{server} components.
\item A Shiny app object created by \code{\link[shiny:shinyApp]{shinyApp()}}.
}}

\item{rawDir}{A directory containing the input raw files,
default is set to the \code{$HOME/Downloads} directory.}

\item{\ldots}{passed to the \code{\link{runApp}} method.}
}
\value{
runs a shiny application
}
\description{
Run the rawDiag shiny application
}
\note{
launch the shiny application by embracing your command line
\itemize{
\item MacOSX and Linux: \verb{R -q -e "rawDiag::shiny(launch.browser = TRUE)"}
\item Microsoft Windows: \verb{R.exe -e "rawDiag::shiny(launch.browser = TRUE)"}
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rawDiag::shiny(rawDir = (rawrr::sampleFilePath() |> dirname()))
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item rawDiag: \doi{10.1021/acs.jproteome.8b00173},
\item rawrr: \doi{10.1021/acs.jproteome.0c00866}
}
}
\author{
Christian Trachsel (2017), Christian Panse (2023)
}
