% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDiag.R
\name{summary.rawDiag}
\alias{summary.rawDiag}
\title{rawDiag Summaries}
\usage{
summary.rawDiag(object)
}
\arguments{
\item{object}{a \code{\link{data.frame}} fullfilling the \code{\link{is.rawDiag}} column naming criteria.}
}
\value{
a table.
}
\description{
rawDiag Summaries
}
